[MISC] move include/seqan3/{core,utility}/simd/detail/default_simd_backend.hpp
