+++ test/unit/core/simd/detail/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_subdirectories()
 
-seqan3_test(default_simd_backend_test.cpp)
 seqan3_test(default_simd_length_builtin_simd_test.cpp)
