+++ test/unit/utility/simd/detail/CMakeLists.txt
@@ -1,2 +1,3 @@
 seqan3_test (builtin_simd_test.cpp)
 seqan3_test (debug_stream_simd_test.cpp)
+seqan3_test (default_simd_backend_test.cpp)
